refactor: organize files in editor folder
